hw-17.2: fix test

+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/grpc/CurrencyGrpcTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Stream;
 
@@ -65,8 +66,8 @@ public class CurrencyGrpcTest extends BaseGrpcTest {
                                 .setAmount(spendSum)
                                 .build()
         );
-        final Double currencyRate = response.getCalculatedAmount();
+        final double currencyRate = response.getCalculatedAmount();
 
-        assertThat(currencyRate).isEqualTo(desiredSum);
+        assertThat(BigDecimal.valueOf(currencyRate)).isEqualTo(BigDecimal.valueOf(desiredSum));
     }
 }
